+++ packages/solidart/lib/src/core/core.dart
@@ -1,5 +1,8 @@
 import 'dart:async';
+import 'dart:collection';
+import 'dart:math';
 
+import 'package:collection/collection.dart';
 import 'package:meta/meta.dart';
 import 'package:solidart/src/utils.dart';
 
@@ -13,3 +16,6 @@ part 'resource.dart';
 part 'signal.dart';
 part 'signal_options.dart';
 part 'signal_base.dart';
+part 'collections/list.dart';
+part 'collections/set.dart';
+part 'collections/map.dart';